requirements.txt
blinker==1.4
celery==4.1.0
Flask==0.12.2
Flask-Migrate==2.1.1
Flask-Redis==0.3.0
Flask-SocketIO==2.9.2
gunicorn==19.7.1
raven==6.3.0
blinker==1.4
Flask_SQLAlchemy==2.3.2
Flask_Migrate==2.1.1
psycopg2==2.7.3.2
Flask_WTF==0.14.2
python_slugify==1.2.4
Flask_Moment==0.5.2
Flask-Login==0.4.1
Flask-WTF==0.14.2